Refactor: Apply MVC pattern to UserDataAllergyVC
- UserDataAllergyViewController에 MVC 패턴 적용
